requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.3.4
redis==5.0.1
requests==2.31.0
facebook-sdk==3.1.0
python-decouple==3.8
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
# Additional dependencies for SSO and OAuth
oauthlib==3.2.2
# For environment configuration
python-dotenv==1.0.0
# For HTML templates
jinja2==3.1.2
# For async file operations
aiofiles==23.2.0